Moved interna/bash to pkg/bash and renamed to pkg/exec.Execute
