+++ internal/cronicle/exec.go
@@ -4,16 +4,16 @@ import (
 	"strings"
 	"time"
 
-	"github.com/jshiv/cronicle/internal/bash"
+	"github.com/jshiv/cronicle/pkg/exec"
 
 	log "github.com/sirupsen/logrus"
 )
 
-//Exec executes task.Command at task.Path and returns the bash.Result struct
+//Exec executes task.Command at task.Path and returns the exec.Result struct
 //prior to execution, the command will replace any ${date}, ${datetime}, ${timestamp}
 //with time t given in the bash command
-func (task *Task) Exec(t time.Time) bash.Result {
-	var result bash.Result
+func (task *Task) Exec(t time.Time) exec.Result {
+	var result exec.Result
 	r := strings.NewReplacer(
 		"${date}", t.Format(TimeArgumentFormatMap["${date}"]),
 		"${datetime}", t.Format(TimeArgumentFormatMap["${datetime}"]),
@@ -26,30 +26,30 @@ func (task *Task) Exec(t time.Time) bash.Result {
 			cmd[i] = s
 		}
 
-		result = bash.Bash(cmd, task.Path)
+		result = exec.Execute(cmd, task.Path)
 	}
 	return result
 }
 
 // Execute does a git pull, git checkout and exec's the given command
-func (task *Task) Execute(t time.Time) (bash.Result, error) {
+func (task *Task) Execute(t time.Time) (exec.Result, error) {
 
 	//Validate the task
 	if err := task.Validate(); err != nil {
-		return bash.Result{}, err
+		return exec.Result{}, err
 	}
 
 	//If a repo is given, clone the repo and task.Git.Open(task.Path)
 	if task.Repo != "" {
 		if err := task.Clone(); err != nil {
-			return bash.Result{}, err
+			return exec.Result{}, err
 		}
 	}
 
 	//Set HEAD and commit state after checkout branch/commit
 	if task.Git.Repository != nil {
 		if err := task.Checkout(); err != nil {
-			return bash.Result{}, err
+			return exec.Result{}, err
 		}
 	}
 
@@ -60,7 +60,7 @@ func (task *Task) Execute(t time.Time) (bash.Result, error) {
 }
 
 //Log logs the exit status, stderr, git commit and other logging data.
-func (task *Task) Log(res bash.Result) {
+func (task *Task) Log(res exec.Result) {
 
 	var commit string
 	var email string
